+++ Wire-iOS/Sources/UserInterface/Helpers/UIColor+Accent.swift
@@ -18,7 +18,54 @@
 
 import Foundation
 
+private var ZM_UNUSED = "UI"
+private var overridenAccentColor: ZMAccentColor = .undefined
+
+
 extension UIColor {
+    
+    
+    /// Set accent color on self user to this index.
+    ///
+    /// - Parameter accentColor: the accent color
+    class func setAccent(_ accentColor: ZMAccentColor) {
+        ZMUserSession.shared()?.enqueueChanges({
+            ZMUser.selfUser()?.accentColorValue = accentColor
+        })
+    }
+    
+    class func accentOverrideColor() -> ZMAccentColor? {
+        return ZMUser.selfUser()?.accentColorValue
+    }
+    
+    class func indexedAccentColor() -> ZMAccentColor {
+        // priority 1: overriden color
+        if overridenAccentColor != .undefined {
+            return overridenAccentColor
+        }
+        
+        guard let activeUserSession = SessionManager.shared?.activeUserSession,
+            ZMUser.selfUser(inUserSession: activeUserSession).accentColorValue != .undefined else {
+                // priority 3: default color
+                return .strongBlue
+        }
+        
+        // priority 2: color from self user
+        return ZMUser.selfUser(inUserSession: activeUserSession).accentColorValue
+    }
+    
+    
+    /// Set override accent color. Can set to ZMAccentColorUndefined to remove override.
+    ///
+    /// - Parameter overrideColor: the override color
+    class func setAccentOverride(_ overrideColor: ZMAccentColor) {
+        if overridenAccentColor == overrideColor {
+            return
+        }
+        
+        overridenAccentColor = overrideColor
+    }
+
     @objc static var accentDarken: UIColor {
         return accent().mix(.black, amount: 0.1).withAlphaComponent(0.32)
     }
